Revert "unlock rtWrapperSceneUpdate when calling js from scene's onUpdate; simplify rtMapObject wrapping (#1355)"
This reverts commit 7c286b2a3a93b66956542b3e91898079b6554e11.
