+++ examples/pxScene2d/src/pxWayland.cpp
@@ -32,10 +32,6 @@
 
 #include "pxContext.h"
 
-#ifdef ENABLE_RT_NODE
-#include "rtScript.h"
-#endif //ENABLE_RT_NODE
-
 #include <map>
 using namespace std;
 
@@ -317,10 +313,6 @@ void pxWayland::onUpdate(double t)
 
   if(!mReadyEmitted && mEvents && mWCtx && (!mUseDispatchThread || !mWaitingForRemoteObject) )
   {
-#ifdef ENABLE_RT_NODE
-    rtWrapperSceneUnlocker unlocker;
-#endif //ENABLE_RT_NODE
-
     mReadyEmitted= true;
     mEvents->isReady(true);
   }
