+++ src/rtScriptNode/rtWrapperUtils.h
@@ -222,5 +222,27 @@ public:
 
 } // namespace
 
+class rtWrapperSceneUnlocker
+{
+public:
+  rtWrapperSceneUnlocker()
+    : m_hadLock(false)
+  {
+    if (rtWrapperSceneUpdateHasLock())
+    {
+      m_hadLock = true;
+      rtWrapperSceneUpdateExit();
+    }
+  }
+
+  ~rtWrapperSceneUnlocker()
+  {
+    if (m_hadLock)
+      rtWrapperSceneUpdateEnter();
+  }
+private:
+  bool m_hadLock;
+};
+
 #endif
